User-management-minikube-postgres/requirements.txt
flask
psycopg2-binary
pydantic